fix: Refactor to useRenderedCount
